Reorg templates
