+++ packages/mux-uploader/src/templates/mux-uploader-file-select.html
