+++ packages/mux-uploader/test/web-test-runner.config.mjs
@@ -23,7 +23,7 @@ export default {
     esbuildPlugin({
       ts: true,
       json: true,
-      loaders: { '.css': 'text', '.svg': 'text' },
+      loaders: { '.css': 'text', '.svg': 'text', '.html': 'text' },
     }),
   ],
   coverageConfig: {
